# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Q-Wash API
  description: >
    Car-washing queue API: phone + SMS-OTP auth, washing point / service
    CRUD, capacity-aware booking availability, a live queue board, and
    booking history. See `docs/API.md` in the repo for the full prose
    contract and `docs/DATA_MODEL.md` for the schema and algorithms this
    spec is generated alongside.
  version: "1.0.0"
servers:
  - url: /
    description: Same-origin (this server also hosts the docs page)
...
security: []
paths:
...
  /api/v1/washing-points/{id}/availability:
    parameters:
      - $ref: "#/components/parameters/washingPointId"
    get:
      tags: [Washing points]
      operationId: getAvailability
      summary: Bookable time windows for a service on a date
      description: >
        Public. Stepped every 15 minutes across the requested date's
        resolved per-weekday schedule window(s) (see GET/PUT
        .../schedule) — two disjoint windows on a day with a break, none
        on a closed day. A slot appears only if the service's
        duration_minutes fits before the window's close and at least one
        box is free for the whole [start, end); available_boxes lists
        which ones, for the client to offer a box choice. Timestamps are
        anchored to the fixed Asia/Dushanbe business timezone, not UTC.
      parameters:
        - name: service_id
          in: query
          required: true
          schema: { type: string, format: uuid }
        - name: date
          in: query
          required: true
          schema: { type: string, format: date, example: "2026-08-06" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AvailabilityList" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/v1/washing-points/{id}/queue:
    parameters:
      - $ref: "#/components/parameters/washingPointId"
    get:
      tags: [Queue]
      operationId: getQueueBoard
      summary: Live queue board for a washing point
      description: >
        Requires role `staff` or `admin`, scoped to their own washing
        point for staff (404 `washing_point_not_found` for another point).
        Bookings with status `queue`/`waiting`/`washing`, ordered by
        `scheduled_start_at`. A separate, display-oriented shape from
        `Booking` — built for a shared/public-facing screen, so it omits
        `user_id`/`car_id` and never exposes a full phone number (only the
        last 4 digits).
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BoardItemList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /api/v1/queue:
    get:
      tags: [Queue]
      operationId: listQueueNetworkWide
      summary: Network-wide live queue board
      description: >
        Requires role `staff` or `admin`. Same live-board shape as
        `GET /washing-points/{id}/queue` but not scoped to a path-level id:
        `admin` may pass `washing_point_id` to scope it, or omit it for
        every point at once; `staff`/`worker` are always forced to their
        own washing point regardless of the query param.
      security: [{ bearerAuth: [] }]
      parameters:
        - name: washing_point_id
          in: query
          required: false
          description: Admin only — ignored (forced to the caller's own point) for staff/worker.
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BoardItemList" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [Queue]
      operationId: createBooking
      summary: Book a queue slot
      description: >
        Requires auth (any role — not restricted to "customer"; a
        staff/admin user can book for themselves too). `washing_point_id`
        is not in the body — derived from `service_id`. `box_number` is
        client-chosen; the overlap check happens inside a transaction that
        locks the washing point row, so concurrent bookings for the same
        point are fully serialized; 409 `slot_unavailable` if the requested
        box isn't free for some instant in the requested interval. A user
        may have at most one active (`queue`/`waiting`/`washing`) booking
        at a time — 409 `active_booking_exists` otherwise.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BookingCreate" }
      responses:
        "201":
          description: Booked
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Booking" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/queue/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Queue]
      operationId: getBooking
      summary: Get a booking
      description: >
        Owner, staff, or admin. Anyone else 404s `queue_not_found` (never
        403 — doesn't confirm the id exists), including staff at a
        different washing point than the booking's own.
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Booking" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/queue/{id}/events:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Queue]
      operationId: streamBooking
      summary: Live-updating stream of a booking (SSE)
      description: >
        Owner only (404, not 403, for a non-owner — same as getBooking).
        text/event-stream: a `Booking` snapshot immediately, then again
        whenever anything changes for its washing point, until the booking
        reaches ready/canceled or the client disconnects. A `: ping`
        comment is sent every 25s to keep the connection alive through
        proxies. Not representable as a normal JSON response — see the
        description for the actual event payload shape (same as Booking).
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: text/event-stream of Booking snapshots
          content:
            text/event-stream:
              schema: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/queue/{id}/cancel:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      tags: [Queue]
      operationId: cancelBooking
      summary: Cancel a booking
      description: >
        Owner-only. Only allowed while status is `queue`/`waiting` — 409
        `cannot_cancel` otherwise, including re-canceling.
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Canceled
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Booking" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/queue/{id}/status:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      tags: [Queue]
      operationId: updateBookingStatus
      summary: Advance a booking's status
      description: >
        Requires role `staff` or `admin`, scoped to the booking's own
        washing point for staff (404 `queue_not_found` otherwise, same as
        a nonexistent id). Forward-only, one step at a time (`queue` →
        `waiting` → `washing` → `ready`); `ready` and `canceled` are
        terminal. 409 `invalid_status_transition` on a skip, a backward
        move, or any move from a terminal state.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BookingStatusUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Booking" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/me/queue:
    get:
      tags: [Queue]
      operationId: getMyBookingHistory
      summary: The caller's full booking history
      description: >
        All statuses, most-recently-scheduled first. Paginated. Only the
        caller's own bookings — no role can see another user's history
        through this endpoint.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/pageSize"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedBookings" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Access token from OTP verify/refresh, e.g. `Authorization: Bearer <token>`"
  parameters:
    washingPointId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    page:
      name: page
      in: query
      description: 1-indexed. Invalid/out-of-range values clamp to the default rather than erroring.
      schema: { type: integer, minimum: 1, default: 1 }
    pageSize:
      name: page_size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing, malformed, or invalid/expired credentials
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Authenticated but not permitted
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Not found (or hidden as not-found to avoid leaking existence)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: State conflict (e.g. slot no longer available, invalid transition)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    BoardItemList:
      type: object
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/BoardItem" } }
      required: [items]
    PaginatedBookings:
      type: object
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/Booking" } }
        page: { type: integer }
        page_size: { type: integer }
        total: { type: integer }
      required: [items, page, page_size, total]
    BookingCreate:
      type: object
      properties:
        car_id: { type: string, format: uuid }
        service_id: { type: string, format: uuid }
        price_option_id: { type: string, format: uuid }
        box_number:
          type: integer
          minimum: 1
          description: >
            Client-chosen box (see the washing point's boxes_count and the
            availability endpoint's available_boxes). The server validates
            it is free for the requested window but never auto-assigns it.
        scheduled_start_at: { type: string, format: date-time }
        notes: { type: string, maxLength: 2000 }
      required: [car_id, service_id, price_option_id, box_number, scheduled_start_at]
    BookingStatusUpdate:
      type: object
      properties:
        status: { type: string, enum: [waiting, washing, ready] }
      required: [status]
...
    AvailabilityList:
      type: object
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/AvailabilitySlot" } }
      required: [items]
...